Reduce contention by make passing "dispatchable" object to DispatchTable.
Vulkan has the concept of "dispatchable" type. An object of a
dispatchable type has a pointer to an internal dispatch table at the
first sizeof(void*) bytes of its memory. This internal dispatch
table is identical for devices/instances and their dispatchable
objects. As an example the pointer is the same for a device and
a command buffer of this device.

This can be used then in the DispatchTable, e.g. we don't need
to look up the device of a command buffer to call some vulkan
functions.
As a result, the bookkeeping can be reduced, which leads to way
less contention on the locks.

Test: Compile